tests for added issue created
